requirements.txt
flask
python-dotenv
openai
flask-socketio
pydub
numpy
